+++ wal-line-project.el
@@ -1,4 +1,4 @@
-;;; wal-line-project.el --- Indicate project -*- lexical-binding: t; -*-
+;;; wal-line-project.el --- Project segment -*- lexical-binding: t; -*-
 
 ;; Author: Krister Schuchardt <krister.schuchardt@gmail.com>
 ;; Homepage: https://github.com/Walheimat/wal-line
@@ -16,6 +16,7 @@
 (require 'wal-line)
 
 (declare-function wal-line--spacer "wal-line.el")
+(declare-function projectile-project-root "ext:projectile.el")
 (declare-function project-root "ext:project.el")
 
 ;;;; Customization:
@@ -33,7 +34,14 @@
 
 (defvar wal-line-project--regexp ".+\\(\\/.+\\)\\/$")
 
-(defun wal-line-project--segment ()
+(defvar-local wal-line-project--segment nil)
+
+(defun wal-line-project--set-segment ()
+  "Set the segment."
+  (when-let ((info (wal-line-project--get-info)))
+    (setq-local wal-line-project--segment (concat (wal-line--spacer) info))))
+
+(defun wal-line-project--get-info ()
   "Get the project or root segment."
   (let ((p-root (cond
                  ((eq wal-line-project-provider 'projectile)
@@ -41,18 +49,25 @@
                  ((eq wal-line-project-provider 'project)
                   (project-root (project-current)))
                  (t nil))))
-    (if (and p-root
-             (buffer-file-name)
-             (string-match-p wal-line-project--regexp p-root))
-        (progn
-          (string-match wal-line-project--regexp p-root)
-          (concat
-           (wal-line--spacer)
-           (propertize (substring (match-string 1 p-root) 1) 'face 'wal-line-emphasis)))
-      "")))
+    (when (and p-root
+               (buffer-file-name)
+               (string-match-p wal-line-project--regexp p-root))
+      (string-match wal-line-project--regexp p-root)
+      (propertize (substring (match-string 1 p-root) 1) 'face 'wal-line-emphasis))))
 
 (defvar wal-line--segments)
 (wal-line-add-segment project)
+
+(defun wal-line-project--setup ()
+  "Set up project segment."
+  (add-hook 'find-file-hook #'wal-line-project--set-segment))
+
+(defun wal-line-project--teardown ()
+  "Set up project segment."
+  (remove-hook 'find-file-hook #'wal-line-project--set-segment))
+
+(add-hook 'wal-line-setup-hook #'wal-line-project--setup)
+(add-hook 'wal-line-teardown-hook #'wal-line-project--teardown)
 
 (provide 'wal-line-project)