+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Position segment shows column number.
 - Version control segment code was refactored to make augmenting the
   segment easier.
+- Same was done for the project segment.
 
 ### Fixed